feat(spinner): implement SpinnerManager for UI spinner functionality
